refactor: updating naming
