+++ packages/odyssey-react-mui/src/labs/DataView/DetailPanel.tsx
