+++ packages/odyssey-react-mui/src/labs/index.ts
@@ -17,7 +17,7 @@ export type { LocalizationProviderProps } from "@mui/x-date-pickers";
 export * from "./DatePicker";
 export * from "./datePickerTheme";
 
-export * from "./DataComponents";
+export * from "./DataView";
 
 /** @deprecated Will be removed in a future Odyssey version in lieu of the one shipping with DataTable */
 export * from "./DataTablePagination";
